Create BMI Calculator
Write a program/class (which you must call BMI) which computes the Body Mass Index (BMI) of a person.  The BMI is computed by dividing the weight of a person (in kgs) by the square of the height of the person in meters.

Notes:
Your program will have input, computation and output.
You can use either the command line or dialog boxes for input and output.
The format and output of your input and output MUST be the same as shown in the Sample Input/Output.  This is important as it will affect the automatic part of the grading of your work.
Sample expected input/output

What is your weight in kg? 80
What is your height in metres? 1.8
Your BMI is 24.691358024691358