Create Loan Repayment Calculator
Write a program/class in Java (which you must call LoanRepaymentCalculator) to compute what the monthly repayment on a loan would be according to the following formula.

 Loan repayment formula
where M is the monthly repayment amount (e.g. 559.55), P is the principal amount of the loan (e.g. 100000), r is the monthly interest rate (e.g. 0.06) and n is the term of the loan in months (e.g. 360).

You must obtain the loan amount, the yearly interest rate and the number of years of the loan (IN THAT ORDER) from the user at the Console.  The monthly interest rate is calculated by dividing the annual interest rate by 12.  The number of months of the loan is calculated by multiplying the number of years by 12.  A sample of the expected input/output is shown below.

Sample expected input/output

Enter loan amount? 100000
Enter annual interest rate (e.g. 0.04)? 0.06
Enter the term of the loan in years? 30
The monthly repayment for a 30 year loan of 100000.00 at an annual interest rate of 0.06 would be 599.55