Create Total Average Calculator
Write a program in Java (and call it TotalAverageCalculator) which uses a for loop to read in between 2 & 10 of integers from the user and which computes the total and average (to two decimal places) of all the numbers entered.  For testing purposes your input and output MUST look like those which are in the Sample input/output.

Sample expected input/output (3 test runs shown)

How many integers do you want to enter? 4
Enter integer 1: 3
Enter integer 2: 7
Enter integer 3: 6
Enter integer 4: 3
The sum of your integers is 19 and the average is 4.75 

How many integers do you want to enter? 2
Enter integer 1: 3
Enter integer 2: 2
The sum of your integers is 5 and the average is 2.50 

How many integers do you want to enter? 45
Error:  This program is constrained to only compute the total & average of between 2 & 10 integers.