Create Incremental Statistics
Using a while or do-while loop, write a program/class (which you must call IncrementalStatistics) which takes keeps taking in whole numbers from a user on the command line until the user enters "quit" or "exit".  The user should be allowed to enter only one number each time and the system should compute the average and variance of all numbers entered so far each time.  You must do this WITHOUT storing all the numbers entered. You must use the formulas which follow which compute the average and variance based only on the previous values (of the average and variance) together with the latest number entered.  The average and variance must be displayed correct to one decimal place.  You must use the input & output text as shown in the sample expected input/output (for automatic testing purposes).

 
Sample expected input/output
This program computes the average and variance of all numbers entered.
Enter a number (or type 'exit'): 1
So far the average is 1.0 and the variance is 0.0
Enter another number (or type 'exit'): 2
So far the average is 1.5 and the variance is 0.3
Enter another number (or type 'exit'): ten
Error:  You must enter a real number (e.g. 12.5)
Enter another number (or type 'exit'): 3
So far the average is 2.0 and the variance is 0.7
Enter another number (or type 'exit'): 10
So far the average is 4.0 and the variance is 12.5
Enter another number (or type 'exit'): 5
So far the average is 4.2 and the variance is 10.2
Enter another number (or type 'exit'): exit
Goodbye.