Create Three Numbers
Write a program (and call the class ThreeNumbers, and the file ThreeNumbers.java) to repeatedly take in three numbers and to report on the median (or middle) value and on how many of the three numbers are greater than the average of the numbers.  As part of your program you must write and make use of the following functions:
A medianOf function which takes three integers and returns the median of the three numbers (i.e. the middle number if considered in order).
An averageOf function which takes three integers and returns the average of the three numbers as a real number (a double).
A countOfNumbersGreaterThanTheAverage function which takes three integers and returns how many of the three numbers are greater than the average (of the three numbers).
A getFormattedOutputString function which takes the median and how many of the numbers are greater than the average and returns a String in the form "The median of your numbers is <median> , and <countOfGreaterThanAverage> of them are greater than their average." where <median>  and  <countOfGreaterThanAverage > are replaced by the passed values.
Sample expected input/output
Please enter your three integers separated by spaces (or enter 'quit'):  2 3 1
The median of your numbers is 2, and 1 of them is greater than their average.
Please enter your three integers separated by spaces (or enter 'quit'):  35 34 23
The median of your numbers is 34, and 2 of them are greater than their average.
Please enter your three integers separated by spaces (or enter 'quit'):  234 234 234
The median of your numbers is 234, and 0 of them is greater than their average.
Please enter your three integers separated by spaces (or enter 'quit'):  quit