Create Pernicious Numbers
Write a program (and call the class PerniciousNumbers, and the file  PerniciousNumbers.java) to output details of all pernicious numbers greater than 0 up to a user specified limit.  For each pernicious number the program should provide output similar to this:  "21 is a pernicious number as it contains 3 ones in it's binary representation (10101)".  Note that a number is called pernicious if it contains a prime number of ones in its binary represenation.  Some other examples of pernicious numbers are 9 (1001 which has 2 ones and 2 is a prime number), 31 (11111 which has 5 ones and 5 is a prime number). 



As part of your program you must write and make use of the following functions:
A getBinaryString function which takes a single integer number and returns a String containing a binary representation of the number.  e.g. if passed 21 it would return "10101".
An isPernicious function which takes a single integer number and returns whether the number is pernicious or not (i.e. whether the number contains a prime number of ones in its binary represenation)
A countBinaryOnes function which takes a single integer number and returns the number of ones in a binary representation of the number.
An isPrime function which takes a single integer number and returns whether it is a prime number or not.

Sample expected input/output
Enter the maximum number you want to consider: 21
3 is a pernicious number as it contains 2 ones in it's binary representation (11)
5 is a pernicious number as it contains 2 ones in it's binary representation (101)
6 is a pernicious number as it contains 2 ones in it's binary representation (110)
7 is a pernicious number as it contains 3 ones in it's binary representation (111)
9 is a pernicious number as it contains 2 ones in it's binary representation (1001)
10 is a pernicious number as it contains 2 ones in it's binary representation (1010)
11 is a pernicious number as it contains 3 ones in it's binary representation (1011)
12 is a pernicious number as it contains 2 ones in it's binary representation (1100)
13 is a pernicious number as it contains 3 ones in it's binary representation (1101)
14 is a pernicious number as it contains 3 ones in it's binary representation (1110)
17 is a pernicious number as it contains 2 ones in it's binary representation (10001)
18 is a pernicious number as it contains 2 ones in it's binary representation (10010)
19 is a pernicious number as it contains 3 ones in it's binary representation (10011)
20 is a pernicious number as it contains 2 ones in it's binary representation (10100)
21 is a pernicious number as it contains 3 ones in it's binary representation (10101)