Create Median Assignment
Write a program (and call the class Median, and the file  Median .java) to take repeatedly take in real numbers from a user and to compute (and report) the median of all numbers entered, the average of all numbers entered and the average of the last 3 numbers entered (or the average of all numbers if there are less than 3 numbers).
As part of your program you must write and make use of the following functions:
A computeMedian function which takes an array of doubles and returns the median value.  The median value is the middle value if the numbers were in an ordered list.  If there are an even number of numbers then the median is the average of the two middle values.   For example if passed { 3.0, 2.8, 4.7, 1.0, 4.0 } the median would be 3.0.  If passed { 3.0, 2.8, 4.7, 1.0 } the median would be 2.9.  If there are no numbers in the array the function returns 0.
A createSortedArray function which takes an array of doubles and returns a new sorted version of the array.    For example if passed { 3.0, 2.8, 4.7, 1.0, 4.0 } it would return { 1.0, 2.8, 3.0, 4.0, 4.7 }.
A computeRollingAverage function which takes an array of doubles (first parameter) and the number of values (at the end of the array) to consider when computing the average.  For example if passed { 3.0, 2.0, 5.0, 1.0, 4.0 } and 3 it would return  3.333333 which is (5.0+1.0+4.0)/3.  If the length of the array is less than the number of values to consider then the average of all values in the array is returned.   For example if passed { 3.0, 6.0 } and 4 it would return 4.5 which is (3.0+6.0)/2.
A convertToString function which takes an array of doubles and returns a formatted String containing the array.   For example if passed { 3.0, 2.79, 4.7, 1.04, 4.0 } it would return "{ 3.0, 2.8, 4.7, 1.0, 4.0 }".  If the passed array is empty it should return "{ }".  All doubles should be output correct to 1 decimal place.
Note that you will need to study the Randomising and Sorting array prerecorded lecture which is provided this week in order to solve part of this problem.  Also note that the String.format("%.2f", number) method might be useful to you (have a look online).
Sample expected input/output
Welcome to the median & rolling average system.
Enter a number (or enter quit): 5.0
The median of { 5.0 } is 5.0 and the rolling average of the last 3 values is 5.0.
Enter a number (or enter quit): 3.0
The median of { 5.0, 3.0 } is 4.0 and the rolling average of the last 3 values is 4.0.
Enter a number (or enter quit): 7
The median of { 5.0, 3.0, 7.0 } is 5.0 and the rolling average of the last 3 values is 5.0.
Enter a number (or enter quit): 99.0
The median of { 5.0, 3.0, 7.0, 99.0 } is 6.0 and the rolling average of the last 3 values is 36.3.
Enter a number (or enter quit): sgd
Error - Enter any real number or quit.
Enter a number (or enter quit): 8
The median of { 5.0, 3.0, 7.0, 99.0, 8.0 } is 7.0 and the rolling average of the last 3 values is 38.0.
Enter a number (or enter quit): quit