Create Grade Computation
Write a program (and call the class GradeComputation, and the file  GradeComputation .java) to determine the end of year grade for students.  The program must repeatedly take all of the required percentages from the user (for each student), and compute the overall grade attained.   The following rules apply:
To pass the year students must obtain an rounded overall weighted average greater than or equal to 40% and must fail no more than 10 credits with no (rounded) mark for any module less than 35%.  If the student has not passed they should be given a "FAIL" grade.
The weights for the rounded overall weighted average are the credits associated with the various modules.
If a student has passed the year, they should be given a grade based on their rounded overall weighted average as follows:
Grade	Range of percentages
I	70-100
II.1	60-69
II.2	50-59
III	40-49
As part of your program you must write and make use of the following functions:
int creditsBelowSpecifiedMark( double[] marks, int specifiedMinimumMark ) which takes an array of student marks and returns the total number of credits associated with any (rounded) marks which are below the specified minimum mark.
double weightedAverageMark( double[] marks ) which takes an array of student marks and returns the weighted average mark (where the weight for each mark is determined by the credits associated with it).
String determineOverallGrade( double[] marks ) which takes an array of student marks and returns the overall grade as a String (One of "FAIL", "III", "II.2", "II.1", "I").
To make your life easier the definition of two constant arrays are provided:
 public static final String[] MODULE_CODES   = { "CSU11001", "CSU11010", "CSU11013", "CSU11021", "CSU11022", "CSU11026", "CSU11031", "CSU11081", "CSU12002", "STU11002" };
 public static final int[]    MODULE_CREDITS = {  5,          10,         5,          5,          5,          10,         5,          5,          5,          5         };

Your program must produce input & output in the form shown in sample input/output below:

Sample expected input/output
Welcome to the first year grade assessor.
Enter the student mark for CSU11001 (or enter quit): 67
Enter the student mark for CSU11010 (or enter quit): oops
Error - Enter a number between 0.0 and 100.0 or quit.
Enter the student mark for CSU11010 (or enter quit): 78
Enter the student mark for CSU11013 (or enter quit): 56.2
Enter the student mark for CSU11021 (or enter quit): 79
Enter the student mark for CSU11022 (or enter quit): 67
Enter the student mark for CSU11026 (or enter quit): 68.5
Enter the student mark for CSU11031 (or enter quit): 70
Enter the student mark for CSU11081 (or enter quit): 72.3
Enter the student mark for CSU12002 (or enter quit): 71
Enter the student mark for STU11002 (or enter quit): 89
Result = I with an overall mark of 72%.

Enter the student mark for CSU11001 (or enter quit): 56
Enter the student mark for CSU11010 (or enter quit): 45
Enter the student mark for CSU11013 (or enter quit): 57
Enter the student mark for CSU11021 (or enter quit): 32
Enter the student mark for CSU11022 (or enter quit): 46
Enter the student mark for CSU11026 (or enter quit): 56
Enter the student mark for CSU11031 (or enter quit): 67
Enter the student mark for CSU11081 (or enter quit): 45
Enter the student mark for CSU12002 (or enter quit): 58
Enter the student mark for STU11002 (or enter quit): 56
Result = FAIL with an overall mark of 52%.
   5 credits were failed.

Enter the student mark for CSU11001 (or enter quit):  35.2
Enter the student mark for CSU11010 (or enter quit): 47
Enter the student mark for CSU11013 (or enter quit): 68
Enter the student mark for CSU11021 (or enter quit): 45
Enter the student mark for CSU11022 (or enter quit): 57.5
Enter the student mark for CSU11026 (or enter quit): 47.2
Enter the student mark for CSU11031 (or enter quit): 37.5
Enter the student mark for CSU11081 (or enter quit): 60
Enter the student mark for CSU12002 (or enter quit): 56
Enter the student mark for STU11002 (or enter quit): 49
Result = II.2 with an overall mark of 50%.

Enter the student mark for CSU11001 (or enter quit): 35.0
Enter the student mark for CSU11010 (or enter quit): 39.0
Enter the student mark for CSU11013 (or enter quit): 50
Enter the student mark for CSU11021 (or enter quit): 69
Enter the student mark for CSU11022 (or enter quit): 72
Enter the student mark for CSU11026 (or enter quit): 68
Enter the student mark for CSU11031 (or enter quit): 57
Enter the student mark for CSU11081 (or enter quit): 46
Enter the student mark for CSU12002 (or enter quit): 32
Enter the student mark for STU11002 (or enter quit): 46
Result = FAIL with an overall mark of 52%.
   20 credits were failed.

Enter the student mark for CSU11001 (or enter quit): quit